api/permissions.py
from rest_framework.permissions import BasePermission


class IsOwner(BasePermission):
    """
    Custom permission to only allow owners of an object to edit it.
    """
    def has_object_permission(self, request, view, obj):
        # permissions are only allowed to the owner of the personagem
        return obj.get_owner == request.user
