requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.0.5
django-model-utils==3.1.2
django-polymorphic==2.0.2
django-rest-framework==0.1.0
django-rest-polymorphic==0.1.5
django-rest-swagger==2.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.8.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
openapi-codec==1.3.2
Pillow==5.1.0
psycopg2-binary==2.7.4
PyJWT==1.6.4
pytz==2018.4
requests==2.18.4
simplejson==3.15.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
whitenoise==3.3.1
